requirements.txt
fastapi
uvicorn
supabase
python-dotenv
cloudinary
python-multipart
pydantic
pydantic-core
email-validator
sendgrid
python-http-client
smtplib